Add patch to support namespace user.pax.* on tmpfs, bug #470644
This patch is needed on all Gentoo to preserve end-to-end support
of PaX flags held in the extended attribute namespace user.pax.*
Even non-hardened Gentoo users need this in case they switch from
a gentoo-sources kernel to hardened.  With the XATTR_PAX markings
already in place, they do not need to remake the PaX flags on all
their ELF objects because it will already be done.